feat: 隐藏索引配置导入 (#9995)

+++ bklog/apps/log_search/constants.py
@@ -928,8 +928,8 @@ class FieldDataTypeEnum(ChoicesEnum):
         return {
             "ip": "string",
             "keyword": "string",
-            "integer": "float",
-            "long": "float",
+            "integer": "int",
+            "long": "long",
             "float": "float",
             "double": "float",
             "object": "object",
